implement offline switch
