+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/pigeons/MapInterfaces.kt
@@ -6581,94 +6581,3 @@ interface Cancelable {
     }
   }
 }
-/**
- * Instance that allows connecting or disconnecting the Mapbox stack to the network.
- *
- * Generated interface from Pigeon that represents a handler of messages from Flutter.
- */
-interface OfflineSwitch {
-  /**
-   * Connects or disconnects the Mapbox stack. If set to false, current and new HTTP requests will fail
-   * with HttpRequestErrorType#ConnectionError.
-   *
-   * @param connected Set false to disconnect the Mapbox stack
-   */
-  fun setMapboxStackConnected(connected: Boolean)
-  /**
-   * Provides information if the Mapbox stack is connected or disconnected via OfflineSwitch.
-   *
-   * @return True if the Mapbox stack is disconnected via setMapboxStackConnected(), false otherwise.
-   */
-  fun isMapboxStackConnected(): Boolean
-  /**
-   * Releases the OfflineSwitch singleton instance.
-   *
-   * Users can call this method if they want to do manual cleanup of the resources allocated by Mapbox services.
-   * If the user calls getInstance() after reset, a new instance of the OfflineSwitch singleton will be allocated.
-   */
-  fun reset()
-
-  companion object {
-    /** The codec used by OfflineSwitch. */
-    val codec: MessageCodec<Any?> by lazy {
-      StandardMessageCodec()
-    }
-    /** Sets up an instance of `OfflineSwitch` to handle messages through the `binaryMessenger`. */
-    @Suppress("UNCHECKED_CAST")
-    fun setUp(binaryMessenger: BinaryMessenger, api: OfflineSwitch?, messageChannelSuffix: String = "") {
-      val separatedMessageChannelSuffix = if (messageChannelSuffix.isNotEmpty()) ".$messageChannelSuffix" else ""
-      run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.mapbox_maps_flutter.OfflineSwitch.setMapboxStackConnected$separatedMessageChannelSuffix", codec)
-        if (api != null) {
-          channel.setMessageHandler { message, reply ->
-            val args = message as List<Any?>
-            val connectedArg = args[0] as Boolean
-            var wrapped: List<Any?>
-            try {
-              api.setMapboxStackConnected(connectedArg)
-              wrapped = listOf<Any?>(null)
-            } catch (exception: Throwable) {
-              wrapped = wrapError(exception)
-            }
-            reply.reply(wrapped)
-          }
-        } else {
-          channel.setMessageHandler(null)
-        }
-      }
-      run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.mapbox_maps_flutter.OfflineSwitch.isMapboxStackConnected$separatedMessageChannelSuffix", codec)
-        if (api != null) {
-          channel.setMessageHandler { _, reply ->
-            var wrapped: List<Any?>
-            try {
-              wrapped = listOf<Any?>(api.isMapboxStackConnected())
-            } catch (exception: Throwable) {
-              wrapped = wrapError(exception)
-            }
-            reply.reply(wrapped)
-          }
-        } else {
-          channel.setMessageHandler(null)
-        }
-      }
-      run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.mapbox_maps_flutter.OfflineSwitch.reset$separatedMessageChannelSuffix", codec)
-        if (api != null) {
-          channel.setMessageHandler { _, reply ->
-            var wrapped: List<Any?>
-            try {
-              api.reset()
-              wrapped = listOf<Any?>(null)
-            } catch (exception: Throwable) {
-              wrapped = wrapError(exception)
-            }
-            reply.reply(wrapped)
-          }
-        } else {
-          channel.setMessageHandler(null)
-        }
-      }
-    }
-  }
-}
