+++ ios/Classes/MapboxMapsPlugin.swift
@@ -21,6 +21,7 @@ public class MapboxMapsPlugin: NSObject, FlutterPlugin {
         _SnapshotterInstanceManagerSetup.setUp(binaryMessenger: binaryMessenger, api: snapshotterInstanceManager)
         _OfflineMapInstanceManagerSetup.setUp(binaryMessenger: binaryMessenger, api: offlineMapInstanceManager)
         _TileStoreInstanceManagerSetup.setUp(binaryMessenger: binaryMessenger, api: offlineMapInstanceManager)
+        _OfflineSwitchSetup.setUp(binaryMessenger: binaryMessenger, api: OfflineSwitch.shared)
 
         LoggingController.setup(binaryMessenger)
     }
