+++ lib/mapbox_maps_flutter.dart
@@ -64,3 +64,4 @@ part 'src/map_events.dart';
 part 'src/offline/offline_messenger.dart';
 part 'src/offline/offline_manager.dart';
 part 'src/offline/tile_store.dart';
+part 'src/offline/offline_switch.dart';
